More changes from the checklist
