+++ docs/content/Button.mdx
@@ -1,6 +1,6 @@
 ---
-componentId: button_v2
-title: Button v2
+componentId: button
+title: Button
 status: Alpha
 source: https://github.com/primer/react/tree/main/src/Button2
 storybook: '/react/storybook?path=/story/composite-components-button2'
