+++ docs/content/IconButton.mdx
@@ -6,9 +6,6 @@ storybook: '/react/storybook?path=/story/composite-components-button2'
 description: An accessible button component with no text and only icon.
 ---
 
-import {Props} from '../../src/props'
-import {ComponentChecklist} from '../../src/component-checklist'
-
 ## Usage
 
 ### Installation
