+++ docs/content/LinkButton.mdx
@@ -6,10 +6,6 @@ storybook: '/react/storybook?path=/story/composite-components-button2'
 description: It is a combination of a button and link. Use this to make your link look like a button.
 ---
 
-import {Props} from '../../src/props'
-import {ComponentChecklist} from '../../src/component-checklist'
-import {ArrowRightIcon} from '@primer/octicons-react'
-
 ## Usage
 
 ### Installation
