+++ docs/content/deprecated/Buttons.mdx
@@ -1,7 +1,7 @@
 ---
 componentId: button
 title: Button
-status: Alpha
+status: deprecated
 source: https://github.com/primer/react/blob/main/src/Button
 storybook: '/react/storybook?path=/story/composite-components-button--default-button'
 ---
