+++ src/deprecated/ActionMenu.tsx
@@ -1,7 +1,7 @@
 import {GroupedListProps, List, ListPropsBase} from './ActionList/List'
 import {Item, ItemProps} from './ActionList/Item'
 import {Divider} from './ActionList/Divider'
-import Button, {ButtonProps} from '../Button'
+import Button, {ButtonProps} from './Button'
 import React, {useCallback, useMemo} from 'react'
 import {AnchoredOverlay} from '../AnchoredOverlay'
 import {useProvidedStateOrCreate} from '../hooks/useProvidedStateOrCreate'
